pyproject.toml
[project]
name = "pokealert"
version = "1.1.0"
description = "Ethical TCG stock alert system (Pokémon, One Piece, MetaZoo, Lorcana, Magic, Yu-Gi-Oh)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "httpx[http2]>=0.27.2",
    "h2>=4.1.0",
    "tenacity>=9.0.0",
    "aiohttp>=3.10.10",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "aiosqlite>=0.20.0",
    "structlog>=24.4.0",
    "rich>=13.9.2",
    "typer>=0.12.5",
]

[project.scripts]
pokealert = "pokealert.main:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]